requirements.txt
dlt>=0.5.1
duckdb>=0.9.2
PyYAML>=6.0.1
